moved __tests__ directory out of lib. removed some npm scripts
