+++ __tests__/utils/calendar/calculate-dimensions.js
@@ -1,4 +1,4 @@
-import { calculateDimensions } from '../../../utility/calendar'
+import { calculateDimensions } from 'lib/utility/calendar'
 
 describe('calculateDimensions', () => {
   it('the item is full on a canvas no draging no resizing', () => {
