+++ __tests__/utils/calendar/get-group-orders.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import { getGroupOrders } from '../../../utility/calendar'
+import { getGroupOrders } from 'lib/utility/calendar'
 
 describe('getGroupOrders', () => {
   xit('WRITE UNIT TEST HERE', () => {})
