+++ __tests__/utils/calendar/get-grouped-items.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import { getGroupedItems } from '../../../utility/calendar'
+import { getGroupedItems } from 'lib/utility/calendar'
 
 describe('getGroupedItems', () => {
   xit('WRITE UNIT TEST HERE', () => {})
