+++ __tests__/utils/calendar/iterate-times.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 
-import { iterateTimes } from '../../../utility/calendar'
+import { iterateTimes } from 'lib/utility/calendar'
 
 describe('iterateTimes', () => {
   xit('WRITE TEST HERE :)', () => {})
