+++ __tests__/utils/calendar/no-stack.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import { nostack } from '../../../utility/calendar'
+import { nostack } from 'lib/utility/calendar'
 
 describe('nostack', () => {
   xit('WRITE UNIT TEST HERE', () => {})
