+++ __tests__/utils/calendar/stack.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import { stack } from '../../../utility/calendar'
+import { stack } from 'lib/utility/calendar'
 
 describe('stack', () => {
   xit('WRITE UNIT TEST HERE', () => {})
