+++ __tests__/utils/dom-helpers/get-parent-position.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import { getParentPosition } from '../../../utility/dom-helpers'
+import { getParentPosition } from 'lib/utility/dom-helpers'
 
 describe('getParentPosition', () => {
   xit('WRITE TEST HERE', () => {})
