+++ package.json
@@ -8,13 +8,11 @@
     "build": "npm run build:lib",
     "build:demo": "echo '!!! Building Demo' && cross-env NODE_ENV=production webpack --progress",
     "build:lib": "echo '!!! Building Library' && cross-env NODE_ENV=production babel src --out-dir lib && node-sass src/lib/Timeline.scss lib/lib/Timeline.css && sed -i'.bak' 's/Timeline\\.scss/Timeline\\.css/g' lib/lib/Timeline.js && rm lib/lib/Timeline.js.bak",
-    "jest": "jest",
-    "jest:update": "jest -u",
     "lint": "eslint --ext .js --ext .jsx ./src",
     "lint:fix": "prettier-eslint --parser babylon --write \"src/**/*.js\"",
     "prepublish": "npm run build:lib",
     "start": "webpack-dev-server --hot --host 0.0.0.0 --display-modules",
-    "test": "npm run build:lib && npm run jest && npm run lint"
+    "test": "jest"
   },
   "files": [
     "lib",
@@ -63,6 +61,14 @@
     ],
     "collectCoverageFrom": [
       "src/**/*.{js,jsx}"
+    ],
+    "modulePaths": [
+      "<rootDir>/src/",
+      "<rootDir>/__tests__"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/node_modules",
+      "<rootDir>/__tests__/test-utility"
     ]
   },
   "dependencies": {
