--- spec/spec-helper.js
@@ -1,3 +1,0 @@
-beforeEach(function() {
-
-});
